requirements.txt
amqp==2.5.2
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-celery-results==1.1.2
django-countries-plus==1.2.1
django-extensions==2.2.5
django-filter==2.2.0
django-languages-plus==1.0.0
django-phonenumber-field==4.0.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==1.2.0
kombu==4.6.7
more-itertools==8.0.2
names==0.3.0
nose==1.3.7
password-generator==0.1.0
phonenumbers==8.11.0
Pillow==6.2.1
python-randomword==1.0.3
pytz==2019.3
Random-Word==1.0.4
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
ssh2-python==0.18.0.post1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
